src/mcp_servers/gitlab/server.py
"""GitLab MCP Server Implementation."""

import asyncio
import json
from typing import Any, Dict, Optional
from mcp.server import Server
from mcp.server.stdio import stdio_server
from mcp.types import Tool, TextContent

# For real implementation, uncomment:
# import gitlab

from .tools import GITLAB_TOOLS


class GitLabMCPServer:
    """MCP Server for GitLab integration."""
    
    def __init__(
        self,
        gitlab_url: str = "https://gitlab.com",
        private_token: str = "",
    ):
        self.gitlab_url = gitlab_url
        self.private_token = private_token
        self.server = Server("gitlab-mcp-server")
        self._gitlab_client: Optional[Any] = None
        
        self._setup_handlers()
    
    @property
    def gl(self):
        """Lazy initialization of GitLab client."""
        if self._gitlab_client is None and self.private_token:
            # Uncomment for real implementation:
            # self._gitlab_client = gitlab.Gitlab(
            #     self.gitlab_url,
            #     private_token=self.private_token,
            # )
            pass
        return self._gitlab_client
    
    def _setup_handlers(self):
        """Set up MCP protocol handlers."""
        
        @self.server.list_tools()
        async def list_tools() -> list[Tool]:
            """Return available GitLab tools."""
            return [
                Tool(
                    name=tool["name"],
                    description=tool["description"],
                    inputSchema=tool["input_schema"],
                )
                for tool in GITLAB_TOOLS
            ]
        
        @self.server.call_tool()
        async def call_tool(name: str, arguments: dict) -> list[TextContent]:
            """Execute a GitLab tool."""
            try:
                result = await self._execute_tool(name, arguments)
                return [TextContent(type="text", text=json.dumps(result, indent=2))]
            except Exception as e:
                return [TextContent(type="text", text=f"Error: {str(e)}")]
    
    async def _execute_tool(self, name: str, arguments: dict) -> Dict[str, Any]:
        """Route tool calls to appropriate handlers."""
        handlers = {
            "gitlab_search_code": self._search_code,
            "gitlab_get_file": self._get_file,
            "gitlab_search_merge_requests": self._search_merge_requests,
            "gitlab_get_merge_request": self._get_merge_request,
            "gitlab_get_commits": self._get_commits,
            "gitlab_get_pipeline": self._get_pipeline,
        }
        
        handler = handlers.get(name)
        if not handler:
            raise ValueError(f"Unknown tool: {name}")
        
        return await handler(**arguments)
    
    # ========================================================================
    # Tool Implementations
    # ========================================================================
    
    async def _search_code(
        self,
        query: str,
        project_id: Optional[str] = None,
        file_extension: Optional[str] = None,
        max_results: int = 10,
    ) -> Dict[str, Any]:
        """Search for code across repositories."""
        # Mock implementation
        return {
            "results": [
                {
                    "project": "backend/api-server",
                    "file": "src/db/connection_pool.py",
                    "line": 45,
                    "content": "pool_size = int(os.getenv('DB_POOL_SIZE', 20))",
                    "url": f"{self.gitlab_url}/backend/api-server/-/blob/main/src/db/connection_pool.py#L45",
                },
                {
                    "project": "backend/api-server",
                    "file": "src/db/connection_pool.py",
                    "line": 52,
                    "content": "max_overflow = int(os.getenv('DB_MAX_OVERFLOW', 10))",
                    "url": f"{self.gitlab_url}/backend/api-server/-/blob/main/src/db/connection_pool.py#L52",
                },
                {
                    "project": "backend/reporting-service",
                    "file": "src/queries/orders.py",
                    "line": 23,
                    "content": "SELECT * FROM orders WHERE created_at > :start_date",
                    "url": f"{self.gitlab_url}/backend/reporting-service/-/blob/main/src/queries/orders.py#L23",
                },
            ],
            "total": 3,
            "query": query,
        }
    
    async def _get_file(
        self,
        project_id: str,
        file_path: str,
        ref: str = "main",
    ) -> Dict[str, Any]:
        """Get file contents from a repository."""
        # Mock implementation
        return {
            "project": project_id,
            "file_path": file_path,
            "ref": ref,
            "size": 2048,
            "content": """
import os
from sqlalchemy import create_engine
from sqlalchemy.pool import QueuePool

class DatabasePool:
    \"\"\"Manages database connection pooling.\"\"\"
    
    def __init__(self):
        self.pool_size = int(os.getenv('DB_POOL_SIZE', 20))
        self.max_overflow = int(os.getenv('DB_MAX_OVERFLOW', 10))
        self.pool_timeout = int(os.getenv('DB_POOL_TIMEOUT', 30))
        
        self.engine = create_engine(
            os.getenv('DATABASE_URL'),
            poolclass=QueuePool,
            pool_size=self.pool_size,
            max_overflow=self.max_overflow,
            pool_timeout=self.pool_timeout,
            pool_pre_ping=True,
        )
    
    def get_connection(self):
        return self.engine.connect()
    
    def get_pool_status(self):
        pool = self.engine.pool
        return {
            'size': pool.size(),
            'checked_in': pool.checkedin(),
            'checked_out': pool.checkedout(),
            'overflow': pool.overflow(),
        }
""",
            "encoding": "utf-8",
            "last_commit": {
                "sha": "abc123def",
                "message": "Increase default pool timeout",
                "author": "dev@example.com",
                "date": "2024-01-10T12:00:00Z",
            },
        }
    
    async def _search_merge_requests(
        self,
        project_id: Optional[str] = None,
        state: str = "merged",
        search: Optional[str] = None,
        max_results: int = 10,
    ) -> Dict[str, Any]:
        """Search for merge requests."""
        # Mock implementation
        return {
            "merge_requests": [
                {
                    "iid": 456,
                    "title": "Change reporting job schedule to run during business hours",
                    "project": "backend/reporting-service",
                    "state": "merged",
                    "author": "analyst@example.com",
                    "merged_at": "2024-01-14T09:00:00Z",
                    "url": f"{self.gitlab_url}/backend/reporting-service/-/merge_requests/456",
                },
                {
                    "iid": 455,
                    "title": "Optimize orders query with date index",
                    "project": "backend/reporting-service",
                    "state": "open",
                    "author": "dba@example.com",
                    "created_at": "2024-01-15T11:30:00Z",
                    "url": f"{self.gitlab_url}/backend/reporting-service/-/merge_requests/455",
                },
            ],
            "total": 2,
            "state": state,
        }
    
    async def _get_merge_request(
        self,
        project_id: str,
        mr_iid: int,
        include_changes: bool = False,
    ) -> Dict[str, Any]:
        """Get details of a specific merge request."""
        # Mock implementation
        mr = {
            "iid": mr_iid,
            "title": "Change reporting job schedule to run during business hours",
            "description": """
## Summary
Changed the reporting job schedule from overnight (2 AM) to business hours (10 AM) 
per request from the analytics team.

## Changes
- Updated cron schedule in `reporting-job.yaml`
- Modified job timeout from 2 hours to 1 hour

## Testing
- Tested in staging environment
- Job completed successfully in 45 minutes
            """,
            "project": project_id,
            "state": "merged",
            "author": {"username": "analyst", "email": "analyst@example.com"},
            "merged_by": {"username": "lead", "email": "lead@example.com"},
            "created_at": "2024-01-13T14:00:00Z",
            "merged_at": "2024-01-14T09:00:00Z",
            "labels": ["scheduling", "reporting"],
        }
        
        if include_changes:
            mr["changes"] = [
                {
                    "old_path": "k8s/reporting-job.yaml",
                    "new_path": "k8s/reporting-job.yaml",
                    "diff": """
@@ -5,7 +5,7 @@ metadata:
 spec:
   schedule: 
-    cron: "0 2 * * *"  # 2 AM daily
+    cron: "0 10 * * *"  # 10 AM daily
   jobTemplate:
     spec:
       template:
                    """,
                },
            ]
        
        return mr
    
    async def _get_commits(
        self,
        project_id: str,
        ref: str = "main",
        since: Optional[str] = None,
        path: Optional[str] = None,
        max_results: int = 20,
    ) -> Dict[str, Any]:
        """Get recent commits for a project."""
        # Mock implementation
        return {
            "project": project_id,
            "ref": ref,
            "commits": [
                {
                    "sha": "abc123def456",
                    "message": "Change reporting job schedule to run during business hours",
                    "author": "analyst@example.com",
                    "date": "2024-01-14T08:55:00Z",
                    "files_changed": 1,
                },
                {
                    "sha": "def789ghi012",
                    "message": "Add retry logic to order exports",
                    "author": "dev@example.com",
                    "date": "2024-01-12T16:30:00Z",
                    "files_changed": 2,
                },
            ],
            "total": 2,
        }
    
    async def _get_pipeline(
        self,
        project_id: str,
        ref: Optional[str] = None,
        pipeline_id: Optional[int] = None,
    ) -> Dict[str, Any]:
        """Get pipeline status for a project."""
        # Mock implementation
        return {
            "project": project_id,
            "pipeline": {
                "id": pipeline_id or 12345,
                "ref": ref or "main",
                "status": "success",
                "created_at": "2024-01-14T09:00:00Z",
                "finished_at": "2024-01-14T09:15:00Z",
                "duration": 900,
                "stages": [
                    {"name": "build", "status": "success"},
                    {"name": "test", "status": "success"},
                    {"name": "deploy", "status": "success"},
                ],
                "triggered_by": "analyst@example.com",
            },
        }
    
    async def run(self):
        """Run the MCP server."""
        async with stdio_server() as (read_stream, write_stream):
            await self.server.run(
                read_stream,
                write_stream,
                self.server.create_initialization_options(),
            )


def main():
    """Entry point for the GitLab MCP server."""
    import os
    
    server = GitLabMCPServer(
        gitlab_url=os.getenv("GITLAB_URL", "https://gitlab.com"),
        private_token=os.getenv("GITLAB_TOKEN", ""),
    )
    
    asyncio.run(server.run())


if __name__ == "__main__":
    main()
