pyproject.toml
[project]
name = "ai-sre-agent"
version = "0.1.0"
description = "AI-powered SRE agent for incident management and support triage"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10,<3.13"
authors = [
    { name = "Your Name", email = "you@example.com" }
]
keywords = ["ai", "sre", "mcp", "langchain", "incident-management"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "langchain>=0.3.0",
    "langchain-anthropic>=0.3.0",
    "langchain-community>=0.3.0",
    "mcp>=1.0.0",
    "chromadb>=0.4.0",
    "sentence-transformers>=2.2.0",
    "atlassian-python-api>=3.41.0",
    "python-gitlab>=4.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.25.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "mypy>=1.8.0",
    "ruff>=0.2.0",
    "pre-commit>=3.6.0",
]
webhook = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
]
all = [
    "ai-sre-agent[dev,webhook]",
]

[project.scripts]
sre-agent = "agent.cli:main"
mcp-jira = "mcp_servers.jira.server:main"
mcp-confluence = "mcp_servers.confluence.server:main"
mcp-gitlab = "mcp_servers.gitlab.server:main"

[project.urls]
Homepage = "https://github.com/yourorg/ai-sre-agent"
Documentation = "https://github.com/yourorg/ai-sre-agent#readme"
Repository = "https://github.com/yourorg/ai-sre-agent"
Issues = "https://github.com/yourorg/ai-sre-agent/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/agent", "src/mcp_servers", "src/config"]

[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "/tests",
]

# ============================================================================
# UV Configuration
# ============================================================================
[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "mypy>=1.8.0",
    "ruff>=0.2.0",
    "pre-commit>=3.6.0",
]

# ============================================================================
# Ruff Configuration (replaces black, isort, flake8)
# ============================================================================
[tool.ruff]
target-version = "py310"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
    "TCH",    # flake8-type-checking
    "PTH",    # flake8-use-pathlib
    "ERA",    # eradicate (commented out code)
    "RUF",    # Ruff-specific rules
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
    "B904",   # raise without from inside except
    "ARG001", # unused function argument
]

[tool.ruff.lint.isort]
known-first-party = ["agent", "mcp_servers", "config"]

[tool.ruff.lint.per-file-ignores]
"tests/**/*" = ["ARG", "S101"]
"**/__init__.py" = ["F401"]

# ============================================================================
# MyPy Configuration
# ============================================================================
[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_configs = true
show_error_codes = true
show_column_numbers = true

[[tool.mypy.overrides]]
module = [
    "langchain.*",
    "langchain_anthropic.*",
    "langchain_community.*",
    "chromadb.*",
    "sentence_transformers.*",
    "atlassian.*",
    "gitlab.*",
    "mcp.*",
]
ignore_missing_imports = true

# ============================================================================
# Pytest Configuration
# ============================================================================
[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
python_files = ["test_*.py", "*_test.py"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--tb=short",
    "--strict-markers",
    "-ra",
]
markers = [
    "slow: marks tests as slow",
    "integration: marks tests as integration tests",
]
filterwarnings = [
    "ignore::DeprecationWarning",
]

[tool.coverage.run]
source = ["src"]
branch = true
omit = ["*/tests/*", "*/__init__.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
]
